chore(unit-tests): fix unit test
OAuth2SummitOrders, MediaUpload, Notifications, OrderExtraQuestion, PresentationAction
and PresentationActionType tests

chore(ci): add OAuth2SummitMediaFileTypeApiControllerTest, OAuth2SummitMediaUploadTypeApiControllerTest,
OAuth2SummitMetricsApiControllerTest, OAuth2SummitNotificationsApiControllerTest,
OAuth2SummitOrderExtraQuestionTypeApiTest, OAuth2SummitOrdersApiTest,
OAuth2SummitPresentationActionApiTest, OAuth2SummitPresentationActionTypeApiTest

+++ tests/InsertSummitTestData.php
@@ -1008,7 +1008,9 @@ trait InsertSummitTestData
         self::$em->remove(self::$summit);
         self::$em->remove(self::$summit2);
         self::$em->remove(self::$summit_permission_group);
-        self::$em->remove(self::$default_media_file_type);
+        if (!is_null(self::$default_media_file_type)) {
+            self::$em->remove(self::$default_media_file_type);
+        }
         self::$em->flush();
 
         // reset static vars
